Added base discriptions and intentions to Read me
still more to add,  Discribes base concept of the game, and purpose of
the exersise but not the games limits, scope, or whats currently in
place.

do note the base capabilitys of the stat game are already in play. one
by buy workers with score, build farms, and assign works, and combat
worker consumption.